Move frames_to_video_bytes to handler.py
